# mint-ensemble-manager/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mint Ensemble Manager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://portals.tapis.io/v3/oauth2/authorize",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
...
      "ModelThread": {
        "description": "",
        "required": [
          "thread_id"
        ],
        "type": "object",
        "properties": {
          "thread_id": {
            "description": "The Modeling thread id",
            "type": "string"
          },
          "model_id": {
            "description": "The Model id (optional. Set to null to run all)",
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/monitors": {
      "post": {
        "summary": "Submit modeling thread for monitoring",
        "operationId": "submitMonitor",
        "tags": [
          "Monitors"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "requestBody": {
          "description": "Modeling thread scenario/subgoal/id",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelThread"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Monitoring thread submitted successfully"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Fetch execution status of modeling thread",
        "operationId": "fetchRunStatus",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the scenario"
          },
          {
            "in": "query",
            "name": "thread_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the thread to check status for"
          }
        ],
        "responses": {
          "200": {
            "description": "Thread Details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Thread details fetched successfully"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}